# Re-implement the user_data heap