--- a/bench/.ocamlformat
+++ b/bench/.ocamlformat
@@ -1,0 +1,4 @@
+enable
+profile = conventional
+break-infix = fit-or-vertical
+module-item-spacing = compact