--- a/bench/main.ml
+++ b/bench/main.ml
@@ -1,7 +1,7 @@
 open Bechamel
 
 let noop_run queue_depth =
-  let t = Uring.create ~queue_depth ~default:(-1) () in
+  let t = Uring.create ~queue_depth () in
   Staged.stage (fun () ->
       for i = 1 to queue_depth do
         let r = Uring.noop t i in
@@ -10,7 +10,7 @@ let noop_run queue_depth =
       let submitted = Uring.submit t in
       assert (submitted = queue_depth);
       for _ = 1 to queue_depth do
-        ignore (Uring.wait t : _ option)
+        ignore (Uring.wait t : _ Uring.completion_option)
       done)
 
 let suite =