--- a/lib/uring/heap.ml
+++ b/lib/uring/heap.ml
@@ -2,77 +2,75 @@ let ( = ) : int -> int -> bool = ( = )
 let ( <> ) : int -> int -> bool = ( <> )
 
 type ptr = int
-let null_ptr = -1
-let is_null x = x = null_ptr
+let slot_taken = -1
+let free_list_nil = -2
 
 (* Free-list allocator *)
 type 'a t =
   { data: 'a array
   (* Pool of potentially-empty data slots. Invariant: an unfreed pointer [p]
-     into this array is valid if one of the following holds:
-
-     1. [is_null free_head];
-     2. [not (is_null free_head)
-         && free_head <> p
-         && is_null free_tail_relation.(p)].
-
-     Either all slots are occupied (1), or there's at least one free slot and
-     [p] is not one of them (2). *)
+     into this array is valid iff [free_tail_relation.(p) = slot_taken]. *)
   ; mutable free_head: ptr
   ; free_tail_relation: ptr array
-  (* A linked list of free blocks, with [free_head] being the first element and
-     [free_tail_relation] mapping each free slot [p] to the next one (or
-     [null_ptr] if [p] is either not free or is the final slot).
+  (* A linked list of pointers to free slots, with [free_head] being the first
+     element and [free_tail_relation] mapping each free slot to the next one.
+     Each entry [x] signals a state of the corresponding [data.(x)] slot:
 
-     Invariant: for all [p : ptr], either [is_null p] or [0 <= p < length]. The
-     user sees non-null pointers only. *)
-  ; max_size: int
+     - [x = slot_taken]: the data slot is taken;
+     - [x = free_list_nil]: the data slot is free, and is last to be allocated;
+     - [0 <= x < length]: the data slot is free, and will be allocated before
+       [free_tail_relation.(x)].
+
+     The user is given only pointers [p] such that [free_tail_relation.(p) =
+     slot_taken]. *)
+  ; length: int
   }
 
 let create : type a. int -> a t =
  fun n ->
   if n < 0 || n > Sys.max_array_length then invalid_arg "Heap.create" ;
   (* Every slot is free, and all but the last have a free successor. *)
-  let free_head = if n = 0 then null_ptr else 0 in
+  let free_head = if n = 0 then free_list_nil else 0 in
   let free_tail_relation = Array.init n succ in
-  if n > 0 then free_tail_relation.(n - 1) <- null_ptr;
+  if n > 0 then free_tail_relation.(n - 1) <- free_list_nil;
   let data =
-    (* See invariants on {!t}, which ensure this data is never read. *)
+    (* No slot in [free_tail_relation] is [slot_taken], so initial data is
+       inaccessible. *)
     Array.make n (Obj.magic `invalid : a)
   in
-  { data; free_head; free_tail_relation; max_size = n }
+  { data; free_head; free_tail_relation; length = n }
 
 exception No_space
 
 let alloc t a =
   let ptr = t.free_head in
-  if is_null ptr then raise No_space;
-  let next_head = t.free_tail_relation.(ptr) in
-  Array.unsafe_set t.free_tail_relation ptr null_ptr;
+  if ptr = free_list_nil then raise No_space;
   Array.unsafe_set t.data ptr a;
-  t.free_head <- next_head;
+
+  (* Drop [ptr] from the free list. *)
+  let tail = Array.unsafe_get t.free_tail_relation ptr in
+  Array.unsafe_set t.free_tail_relation ptr slot_taken;
+  t.free_head <- tail;
+
   ptr
 
 let free : type a. a t -> ptr -> a =
  fun t ptr ->
-  let () =
-    assert (not (is_null ptr)) (* [alloc] returns only valid pointers. *);
-    if ptr >= t.max_size then invalid_arg "Heap.free: invalid pointer";
-    let next_free = Array.unsafe_get t.free_tail_relation ptr in
+  assert (ptr >= 0) (* [alloc] returns only valid pointers. *);
+  if ptr >= t.length then Fmt.invalid_arg "Heap.free: invalid pointer %d" ptr;
+  let slot_state = Array.unsafe_get t.free_tail_relation ptr in
+  if slot_state <> slot_taken then invalid_arg "Heap.free: pointer already freed";
 
-    if is_null t.free_head then
-      (* No free slots, so [ptr] is valid in [t.data]. Case (1) above. *)
-      assert (is_null next_free)
-
-    else if t.free_head <> ptr && is_null next_free then
-      (* One or more free slots, and [ptr] isn't one of them. Case (2) above. *)
-      Array.unsafe_set t.free_tail_relation ptr t.free_head
-
-    else invalid_arg "Heap.free: pointer already freed"
-  in
-  t.free_head <- ptr;
+  (* [t.free_tail_relation.(ptr) = slot_taken], so [t.data.(ptr)] is valid. *)
   let datum = Array.unsafe_get t.data ptr in
-  (* Zero the freed slot. Since this value is now inaccessible (with [ptr] is
-      in the free list) we might as well let it be GCed. *)
+
+  (* Cons [ptr] to the free-list. *)
+  Array.unsafe_set t.free_tail_relation ptr t.free_head;
+  t.free_head <- ptr;
+
+  (* We've marked this slot as free, so [t.data.(ptr)] is inaccessible. We zero
+     it to allow it to be GC-ed. *)
+  assert (t.free_tail_relation.(ptr) <> slot_taken);
   Array.unsafe_set t.data ptr (Obj.magic `invalid : a);
+
   datum