--- a/tests/main.ml
+++ b/tests/main.ml
@@ -8,12 +8,44 @@ let check_string ~__POS__ ~expected = Alcotest.(check ~pos:__POS__ string) "" ex
 module Heap = struct
   module Heap = Uring.Private.Heap
 
-  let test_basic () =
-    let t = Heap.create 10 in
-    let p1 = Heap.alloc t 1 in
-    let p2 = Heap.alloc t 2 in
-    check_int ~__POS__ ~expected:2 (Heap.free t p2);
-    check_int ~__POS__ ~expected:1 (Heap.free t p1)
+  let random_hashtbl_elt tbl =
+    let rec inner n acc (seq : _ Seq.t) =
+      match seq () with
+      | Nil -> acc
+      | Cons (x, xf) ->
+        let acc = if Random.int n = 0 then x else acc in
+        inner (succ n) acc xf
+    in
+    match Hashtbl.to_seq tbl () with
+    | Nil -> invalid_arg "random_hashtbl_elt"
+    | Cons (x, xf) -> inner 1 x xf
+
+  let check_raises_no_space ~__POS__:pos f =
+    Alcotest.check_raises ~pos "" Heap.No_space (fun () -> ignore (f ()))
+
+  let test_normal_usage () =
+    let max_size = 10 in
+    let t = Heap.create max_size in
+    let reference : (Heap.ptr, int) Hashtbl.t = Hashtbl.create max_size in
+    let currently_allocated = ref 0 in
+    for _ = 1 to 100_000 do
+      let attempt_alloc = !currently_allocated = 0 || Random.bool () in
+      match attempt_alloc with
+      | true ->
+        if !currently_allocated = max_size then
+          check_raises_no_space ~__POS__ (fun () -> Heap.alloc t 0)
+        else
+          let data = Random.int 5000 in
+          let ptr = Heap.alloc t data in
+          Hashtbl.replace reference ptr data;
+          incr currently_allocated
+      | false ->
+        let (k, v) = random_hashtbl_elt reference in
+        let v' = Heap.free t k in
+        Hashtbl.remove reference k;
+        check_int ~__POS__ ~expected:v v';
+        decr currently_allocated
+    done
 
   let test_double_free () =
     let () =
@@ -36,9 +68,6 @@ module Heap = struct
     ()
 
   let test_out_of_space () =
-    let check_raises_no_space ~__POS__:pos f =
-      Alcotest.check_raises ~pos "" Heap.No_space (fun () -> ignore (f ()))
-    in
     let () =
       let t = Heap.create 0 in
       (* 1 > 0 *)
@@ -142,10 +171,11 @@ let test_readv () =
 
 let () =
   Test_data.setup ();
+  Random.self_init ();
   let tc name f = Alcotest.test_case name `Quick f in
   Alcotest.run __FILE__ [
     "heap", [
-      tc "" Heap.test_basic;
+      tc "normal_usage" Heap.test_normal_usage;
       tc "double_free" Heap.test_double_free;
       tc "out_of_space" Heap.test_out_of_space;
     ];