In benchmarking Uring + Index, I noticed that large buffers were being held hostage by the `user_data` array, which doesn't drop user data after it has been returned.

This PR replaces the existing free-list allocator with one that doesn't leak memory. Fortunately, the approach used to avoid memory leaks also ensures that it's safe to not require dummy user data when first constructing a uring, so we can get rid of that.