@talex5 2021-04-23
Using an array instead of a linked-list to track the free slots seems reasonable. Would be interesting to see some performance numbers.