@CraigFe 2021-04-23
Microbenchmark:

```ocaml
let noop_run size =
  let t = Uring.create ~queue_depth:size () in
  Staged.stage (fun () ->
      for i = 1 to size do let r = Uring.noop t i in assert r done;
      let submitted = Uring.submit t in assert (submitted = size);
      for _ = 1 to size do ignore (Uring.wait t : (int * int) option) done)
```

On current `main`:

```
╭──────────────┬───────────────────────────┬───────────────────────────┬───────────────────────────╮
│name          │  major-allocated          │  minor-allocated          │  monotonic-clock          │
├──────────────┼───────────────────────────┼───────────────────────────┼───────────────────────────┤
│  noop    10  │             0.0077 mjw/run│           150.0551 mnw/run│           2502.0685 ns/run│
│  noop    30  │             0.0754 mjw/run│           450.1358 mnw/run│           6375.0174 ns/run│
│  noop   100  │             0.8467 mjw/run│          1500.3405 mnw/run│          21088.8327 ns/run│
│  noop   300  │             7.7571 mjw/run│          4500.5976 mnw/run│          63188.8516 ns/run│
│  noop  1000  │            85.7926 mjw/run│         15001.0638 mnw/run│         201523.0778 ns/run│
│  noop  3000  │           794.9051 mjw/run│         45001.8519 mnw/run│         614649.4986 ns/run│
│  noop 10000  │          9265.2980 mjw/run│        150003.3333 mnw/run│        2125486.8748 ns/run│
╰──────────────┴───────────────────────────┴───────────────────────────┴───────────────────────────╯
```

This PR:

```
╭──────────────┬───────────────────────────┬───────────────────────────┬───────────────────────────╮
│name          │  major-allocated          │  minor-allocated          │  monotonic-clock          │
├──────────────┼───────────────────────────┼───────────────────────────┼───────────────────────────┤
│  noop    10  │             0.0000 mjw/run│           120.0567 mnw/run│           2538.0591 ns/run│
│  noop    30  │             0.0000 mjw/run│           360.1439 mnw/run│           6741.1804 ns/run│
│  noop   100  │             0.0000 mjw/run│          1200.3505 mnw/run│          21781.5844 ns/run│
│  noop   300  │             0.0000 mjw/run│          3600.5792 mnw/run│          59570.2872 ns/run│
│  noop  1000  │             0.0000 mjw/run│         12001.0490 mnw/run│         197404.5607 ns/run│
│  noop  3000  │             0.0000 mjw/run│         36001.8072 mnw/run│         581986.1964 ns/run│
│  noop 10000  │             0.0000 mjw/run│        120003.3333 mnw/run│        1971265.5826 ns/run│
╰──────────────┴───────────────────────────┴───────────────────────────┴───────────────────────────╯
```

So it looks like we had almost exactly 15 minor words per noop cycle before, and now we have 12. It's possible to take it down to 8 by inlining the `with_id` in `Uring.noop`, after which I think the remaining allocations are all options / pairs / closures in `Uring.wait`.