@talex5 2021-04-26
Thanks for the benchmark. I tried removing all the `unsafe` and `magic` bits, and it appears to make no difference to the speed:
https://github.com/talex5/ocaml-uring/commit/a240d36f9fe92023b4969b5614b6bab79e55efc4