@talex5 2021-04-26
Nice. Can we get rid of the `magic` too? I know it "should" be impossible for it ever to be returned, but then we thought that about the old code too and you mentioned seeing it return the default value sometimes. It gets pretty hard to debug stuff once the types are wrong (I imagine this could happen if two mutlicore domains tried to the use the same uring, for example).