@CraigFe 2021-04-26
I think it's worth avoiding propagating a `~default` parameter throughout user code: there's nothing they can realistically do when it's returned to them, and returning only valid data is already part of the API contract. Needing to pass this parameter is also a limitation to users: I currently have a "request" type that uses a backpointer into the parent structure, and it's not possible to build one of these backpointers w/o first constructing the Uring.

If you're really set against using `Obj` here, I suggest we put options in the slots and take the performance loss. It's similar to the linked list in terms of perf, but keeps the nicer API and fails more gracefully than `Obj`.